Refactor : 파일 이동 및 파일 경로 수정
폴더 구조에 맞게 파일 수정
