+++ frontend/src/components/drawTools/eraser.tsx
