+++ frontend/src/components/drawTools/palette.tsx
