+++ frontend/src/components/drawTools/pen.tsx
